reorg(0254): extract course dedup to flat _course_dedup.py
build_teaching_yaml.py was dual-role: the Phase-1 teaching-YAML builder, but
also imported by analyze_syllabi.py for its _dedup_course_names helper. That
script→script import blocks the wave-1 move.

Relocate _dedup_course_names plus its OVERLAP_THRESHOLD / MIN_SHARED_READINGS
thresholds verbatim into a neutral flat _course_dedup.py (the 0250 pattern) and
repoint analyze_syllabi.py. The extracted module keeps the "build_teaching_yaml"
logger name so the dedup log line is unchanged. Output byte-identical by
construction (verified: a synthetic frame with co-listed courses merges
identically old vs new). build_teaching_yaml.py's only remaining non-test
importer is gone — it is now import-leaf on the script surface and free to move
in wave-1 (0256). (Its test imports use flat module names, handled by the 0253
path model.)

Ticket-ref: tickets/0254-reorg-code-entry-point-vs-library-classi.erg
