+++ scripts/build_teaching_yaml.py
@@ -22,6 +22,7 @@ from collections import defaultdict
 
 import pandas as pd
 import yaml
+from _course_dedup import _dedup_course_names
 from utils import DATA_DIR, clean_doi, get_logger
 
 log = get_logger("build_teaching_yaml")
@@ -32,8 +33,6 @@ OUTPUT_YAML = os.path.join(DATA_DIR, "teaching_sources.yaml")
 MIN_COURSES = 1  # DOI entries: keep all — DOI from a classified syllabus is reliable
 MIN_COURSES_NO_DOI = 3  # Title-only entries: higher bar (>=3 syllabi)
 MIN_READINGS_DETAILED = 20  # Courses with >=20 DOI readings are "detailed syllabi"
-OVERLAP_THRESHOLD = 0.8  # Course pairs sharing >80% readings are duplicates
-MIN_SHARED_READINGS = 10  # Require >=10 shared readings to consider dedup
 FUZZY_THRESHOLD = 75  # rapidfuzz token_sort_ratio threshold for title grouping
 FUZZY_MIN_WORDS = 4   # titles shorter than this skip fuzzy matching (too generic)
 
@@ -218,63 +217,6 @@ def _fuzzy_dedup_title_only(df):
                  len(nodoi_rows), len(merged_rows))
 
     return result
-
-
-def _dedup_course_names(df):
-    """Merge near-duplicate course names and recompute n_courses.
-
-    Two courses are considered duplicates if they share >=MIN_SHARED_READINGS
-    AND >OVERLAP_THRESHOLD of the smaller course's readings. This catches
-    co-organized MOOCs listed under multiple institution names.
-    """
-    # Build course -> set of row indices
-    course_rows = defaultdict(set)
-    for idx, row in df.iterrows():
-        for c in str(row.get("courses", "")).split(";"):
-            c = c.strip()
-            if c:
-                course_rows[c].add(idx)
-
-    # Find overlapping course pairs
-    courses = list(course_rows.keys())
-    merged = {}  # alias -> canonical
-    for i, c1 in enumerate(courses):
-        if c1 in merged:
-            continue
-        for c2 in courses[i + 1:]:
-            if c2 in merged:
-                continue
-            s1, s2 = course_rows[c1], course_rows[c2]
-            if not s1 or not s2:
-                continue
-            n_shared = len(s1 & s2)
-            overlap = n_shared / min(len(s1), len(s2))
-            if n_shared >= MIN_SHARED_READINGS and overlap > OVERLAP_THRESHOLD:
-                canonical = c1 if len(c1) <= len(c2) else c2
-                alias = c2 if canonical == c1 else c1
-                merged[alias] = canonical
-
-    if not merged:
-        return df
-
-    log.info("  Course dedup: merged %d duplicate course names", len(merged))
-
-    def apply_merge(courses_str):
-        parts = [c.strip() for c in str(courses_str).split(";")]
-        deduped = []
-        seen = set()
-        for c in parts:
-            canonical = merged.get(c, c)
-            if canonical and canonical not in seen:
-                deduped.append(canonical)
-                seen.add(canonical)
-        return " ; ".join(sorted(deduped))
-
-    df = df.copy()
-    df["courses"] = df["courses"].apply(apply_merge)
-    df["n_courses"] = df["courses"].apply(
-        lambda x: len([c for c in x.split(" ; ") if c.strip()]))
-    return df
 
 
 # --- Source 1: scraped readings ---
